build: bundle webview script
