+++ package.json
@@ -118,7 +118,8 @@
         "compile": "tsc -p ./",
         "watch": "tsc -watch -p ./",
         "lint": "eslint src --ext ts",
-        "test": "mocha -r ts-node/register test/**/*.test.ts"
+        "test": "mocha -r ts-node/register test/**/*.test.ts",
+        "build:webview": "ts-node scripts/build-webview.ts"
     },
     "devDependencies": {
         "@types/chai": "^5.2.2",
@@ -126,14 +127,17 @@
         "@types/mock-require": "^3.0.0",
         "@types/moo": "^0.5.10",
         "@types/node": "^14.x.x",
+        "@types/node-fetch": "^2.6.12",
         "@types/vscode": "^1.60.0",
         "@typescript-eslint/eslint-plugin": "^4.x.x",
         "@typescript-eslint/parser": "^4.x.x",
         "@vscode/vsce": "^3.3.2",
         "chai": "^4.3.0",
+        "esbuild": "^0.19.12",
         "eslint": "^7.x.x",
         "mocha": "^10.0.0",
         "mock-require": "^3.0.3",
+        "terser": "^5.41.0",
         "ts-node": "^10.9.1",
         "typescript": "^4.x.x"
     },
